Iteration 4
Added Reservation functionality:
- bookings hold for 2 hours
- changed Log Factory to accommodate for extra reservation logging needs
- table can be booked
- table checks it's status on given day and time to deny/allow booking
- each table keeps record of their reservations (booking time, booking date, who booked and when)
- Added booking test in main

+++ ArchitectureProject/ArchitectureProject/CashPayment.cs
@@ -37,7 +37,7 @@ namespace ArchitectureProject
             }
             else
             {
-                int dif = sum - givenSum;
+                double dif = sum - givenSum;
                 if(dif < 0)
                 {
                     succ = true;
